cryptopay

+++ src/main/java/com/Cryptopay/dtos/TransactionDto.java
@@ -2,12 +2,12 @@ package com.Cryptopay.dtos;
 
 import lombok.*;
 
-import java.math.BigDecimal;
+
 
 @Getter
 @AllArgsConstructor
 @EqualsAndHashCode
 public class TransactionDto {
     private final String subject;
-    private final BigDecimal Amount;
+    private final Double Amount;
 }
